# backend/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Marketplace Order Engine",
    "version": "1.0.0",
    "description": "Demo local: los headers de contexto no constituyen autenticación. Money.amount expresa unidades menores como string."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
    "/quotes": {
      "post": {
        "summary": "Persistir una cotización",
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "in": "header",
            "name": "X-Country",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "in": "header",
            "name": "X-Customer-ID",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Snapshot persistido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quote"
                }
              }
            }
          },
          "400": {
            "description": "Error de negocio o servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Error de negocio o servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Error de negocio o servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error de negocio o servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/orders": {
      "post": {
        "summary": "Confirmar el snapshot una sola vez",
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "in": "header",
            "name": "X-Country",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "in": "header",
            "name": "X-Customer-ID",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pedido confirmado y persistido, o replay idéntico",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Error de negocio o servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Error de negocio o servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Error de negocio o servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error de negocio o servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/quotes/{id}": {
      "get": {
        "summary": "Consultar Quote",
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "in": "header",
            "name": "X-Country",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "in": "header",
            "name": "X-Customer-ID",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Documento del tenant/país/cliente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Quote"
                }
              }
            }
          },
          "400": {
            "description": "Error de negocio o servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Error de negocio o servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Error de negocio o servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error de negocio o servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "summary": "Consultar Order",
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "in": "header",
            "name": "X-Country",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "in": "header",
            "name": "X-Customer-ID",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Documento del tenant/país/cliente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Error de negocio o servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Error de negocio o servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Error de negocio o servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error de negocio o servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "responses": {
          "200": {
            "description": "Disponible",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "status"
                  ],
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/ready": {
      "get": {
        "responses": {
          "200": {
            "description": "Disponible",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "status"
                  ],
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Dependencias no disponibles",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}/details": {
      "get": {
        "summary": "Detalle de la orden, snapshot confirmado y estados demostrativos de ERP/PUSH",
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "in": "header",
            "name": "X-Country",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "in": "header",
            "name": "X-Customer-ID",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Documento del tenant/país/cliente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetails"
                }
              }
            }
          },
          "400": {
            "description": "Error de negocio o servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Error de negocio o servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Error de negocio o servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error de negocio o servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Money": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "amount",
          "currency",
          "scale"
        ],
        "properties": {
          "amount": {
            "type": "string",
            "pattern": "^[0-9]+$",
            "description": "Unidades menores exactas, no unidades mayores decimales."
          },
          "currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$"
          },
          "scale": {
            "type": "integer",
            "minimum": 0,
            "maximum": 6
          }
        }
      },
      "CartItem": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "sku",
          "quantity"
        ],
        "properties": {
          "sku": {
            "type": "string",
            "minLength": 1
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000000
          }
        }
      },
      "QuoteRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "tenantId",
          "country",
          "customerId",
          "paymentMethod",
          "items"
        ],
        "properties": {
          "tenantId": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "paymentMethod": {
            "type": "string",
            "enum": [
              "CREDIT",
              "CASH"
            ]
          },
          "items": {
            "type": "array",
            "minItems": 1,
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/CartItem"
            }
          }
        }
      },
      "OrderRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "quoteId",
          "customerId"
        ],
        "properties": {
          "quoteId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          }
        }
      },
      "Line": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "sku",
          "quantity",
          "comboQuantity",
          "unitPrice",
          "gross",
          "discount",
          "taxableBase",
          "tax"
        ],
        "properties": {
          "sku": {
            "type": "string"
          },
          "quantity": {
            "type": "integer"
          },
          "comboQuantity": {
            "type": "integer"
          },
          "unitPrice": {
            "$ref": "#/components/schemas/Money"
          },
          "gross": {
            "$ref": "#/components/schemas/Money"
          },
          "discount": {
            "$ref": "#/components/schemas/Money"
          },
          "taxableBase": {
            "$ref": "#/components/schemas/Money"
          },
          "tax": {
            "$ref": "#/components/schemas/Money"
          }
        }
      },
      "Adjustment": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "promotionId",
          "promotionType",
          "sku",
          "quantity",
          "discountAmount"
        ],
        "properties": {
          "promotionId": {
            "type": "string"
          },
          "promotionType": {
            "type": "string",
            "enum": [
              "COMBO",
              "SCALE"
            ]
          },
          "sku": {
            "type": "string"
          },
          "quantity": {
            "type": "integer"
          },
          "discountAmount": {
            "$ref": "#/components/schemas/Money"
          }
        }
      },
      "Gift": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "sku",
          "quantity",
          "promotionId",
          "referenceValue",
          "tax"
        ],
        "properties": {
          "sku": {
            "type": "string"
          },
          "quantity": {
            "type": "integer"
          },
          "promotionId": {
            "type": "string"
          },
          "referenceValue": {
            "$ref": "#/components/schemas/Money"
          },
          "tax": {
            "$ref": "#/components/schemas/Money"
          }
        }
      },
      "Quote": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "tenantId",
          "country",
          "customerId",
          "quoteId",
          "currency",
          "paymentMethod",
          "lines",
          "adjustments",
          "gifts",
          "grossSubtotal",
          "discountTotal",
          "taxableBase",
          "taxTotal",
          "total",
          "creditEvaluation",
          "createdAt",
          "expiresAt"
        ],
        "properties": {
          "tenantId": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "quoteId": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "paymentMethod": {
            "type": "string",
            "enum": [
              "CREDIT",
              "CASH"
            ]
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Line"
            }
          },
          "adjustments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Adjustment"
            }
          },
          "gifts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Gift"
            }
          },
          "grossSubtotal": {
            "$ref": "#/components/schemas/Money"
          },
          "discountTotal": {
            "$ref": "#/components/schemas/Money"
          },
          "taxableBase": {
            "$ref": "#/components/schemas/Money"
          },
          "taxTotal": {
            "$ref": "#/components/schemas/Money"
          },
          "total": {
            "$ref": "#/components/schemas/Money"
          },
          "creditEvaluation": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "eligible",
              "available"
            ],
            "properties": {
              "eligible": {
                "type": "boolean"
              },
              "available": {
                "$ref": "#/components/schemas/Money"
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Order": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "tenantId",
          "country",
          "customerId",
          "orderId",
          "orderNumber",
          "quoteId",
          "currency",
          "total",
          "status",
          "createdAt"
        ],
        "properties": {
          "tenantId": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "orderNumber": {
            "type": "string"
          },
          "quoteId": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "total": {
            "$ref": "#/components/schemas/Money"
          },
          "status": {
            "const": "CONFIRMED"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
      "DeliveryState": {
        "type": "object",
        "required": [
          "done",
          "dead",
          "failures",
          "nextAttempt"
        ],
        "properties": {
          "done": {
            "type": "boolean"
          },
          "dead": {
            "type": "boolean"
          },
          "failures": {
            "type": "integer",
            "minimum": 0
          },
          "nextAttempt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "OrderDetails": {
        "type": "object",
        "required": [
          "order",
          "quote",
          "deliveries"
        ],
        "properties": {
          "order": {
            "$ref": "#/components/schemas/Order"
          },
          "quote": {
            "$ref": "#/components/schemas/Quote"
          },
          "deliveries": {
            "type": [
              "object",
              "null"
            ],
            "properties": {
              "erp": {
                "$ref": "#/components/schemas/DeliveryState"
              },
              "push": {
                "$ref": "#/components/schemas/DeliveryState"
              }
            },
            "description": "Estado persistido por destino. null significa sin información disponible; done solo es true tras completar y persistir el efecto."
          }
        }
      }
    }
  }
}